Abe's obsession with the Constitution will not dissolve the cabinet.
and Also, because it is dead? Get out way my building to me can't be said for emperor...Because of the shamelessness of the Government, it will rush through and it is not able to be by kaketsuke-keigo to pass , so  it could be Japanese constitution revised. That's why the emperor of Abe Shinzo Korea.Abe Shinzo and Trump's Rendezvous. Because Trump is Mr. Clinton's conspiracy who says that he no way without me ... Triangle with Masako, Abe and Trump