Clinton... sir.
Clinton is the one who studied with a drill and Prime Minister Abe also enters. President Clinton has come to suspicion that he wulld had kill his daughter. and President Clinton going with Trump president. so  It has nothing to do with the Clinton family. Clinton is the origin that wants to blame Elizabeth. where is relationship Elizabeth with Clinton.